requirements.txt
requests==2.31.0
python-dotenv==1.0.0 
python-telegram-bot[job-queue]==20.7
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9